# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenHall API",
    "description": "Self-hosted school presence and movement platform API.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-openhall_session",
        "description": "Opaque server-side session bearer (development: openhall_session_dev)."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Per-session CSRF token from GET /api/v1/auth/session."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/bootstrap/prepare": {
      "post": {
        "operationId": "prepareBootstrap",
        "tags": [
          "bootstrap"
        ],
        "description": "Validates tenant/school/admin details plus the OIDC provider (live discovery), stores the encrypted draft, and starts the bootstrap OIDC flow. The operator credential travels in the Authorization header, never query or ambient cookies.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tenantName",
                  "tenantSlug",
                  "schoolName",
                  "schoolSlug",
                  "schoolTimeZone",
                  "adminGivenName",
                  "adminFamilyName",
                  "adminDisplayName",
                  "providerKey",
                  "providerDisplayName",
                  "providerIssuer",
                  "providerClientId",
                  "providerClientSecret",
                  "providerAuthMethod",
                  "providerScopes"
                ],
                "properties": {
                  "tenantName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "tenantSlug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 63
                  },
                  "schoolName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "schoolSlug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 63
                  },
                  "schoolTimeZone": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "adminGivenName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "adminFamilyName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "adminDisplayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "providerKey": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 63
                  },
                  "providerDisplayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "providerPreset": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "google"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "generic"
                        ]
                      }
                    ]
                  },
                  "providerIssuer": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "providerClientId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "providerClientSecret": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "providerAuthMethod": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "client_secret_post"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "client_secret_basic"
                        ]
                      }
                    ]
                  },
                  "providerScopes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "minItems": 1,
                    "maxItems": 20
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "authorizationUrl"
                  ],
                  "properties": {
                    "authorizationUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid setup details or provider configuration",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]*$"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Invalid bootstrap token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]*$"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Installation already exists",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]*$"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}